+++ src/main/java/org/scoula/domain/ranking/WeeklyRankingDto.java
@@ -1,4 +1,4 @@
-package org.scoula.domain;
+package org.scoula.domain.ranking;
 
 import lombok.Data;
